refactor usage service
