+++ src/services/transcription.py
@@ -14,7 +14,6 @@ from sqlalchemy.orm import Session
 from sqlalchemy import desc
 
 from src.models.users import User
-from src.models.usage import Usage
 from src.models.transcription import (
   Transcription,
   TranscriptionChunk,
@@ -377,17 +376,3 @@ class TranscriptionService:
     )
     
     return response
-  
-  def get_current_usage(self, session: Session, user_id: str):
-    print("!!!getting usage obj")
-    return session.query(Usage).filter(Usage.user_id == user_id) \
-                                .order_by(desc(Usage.created_at)).first()
-
-  def get_current_usage_quota(self, session: Session, user_id: str):
-    print("!!!getting quota")
-    usage = self.get_current_usage(session, user_id)
-
-    print("!!!lets go")
-
-    # return  {"quota": usage.quota}
-    return usage.quota
