+++ src/services/users.py
@@ -3,6 +3,7 @@ from sqlalchemy.orm import Session
 from src.exceptions.users import InvalidFieldName
 
 from src.models.users import User
+from src.models.usages import Usage
 from src.schemas.auth import RegisterSchema
 from src.utils.db import get_db
 from starlette.status import HTTP_401_UNAUTHORIZED
@@ -11,8 +12,15 @@ from starlette.status import HTTP_401_UNAUTHORIZED
 def create_user(session: Session, user: RegisterSchema):
     db_user = User(**user.model_dump())
     session.add(db_user)
+
+    usage = Usage(user_id=db_user.id)
+    session.add(usage)
+
     session.commit()
+
     session.refresh(db_user)
+    session.refresh(usage)
+    
     return db_user
